+++ src/main/java/ke/co/miles/kcep/mis/entities/InputType.java
@@ -41,7 +41,7 @@ public class InputType implements Serializable {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Basic(optional = false)
     @Column(name = "id")
-    private Integer id;
+    private Short id;
     @Size(max = 45)
     @Column(name = "type")
     private String type;
@@ -54,15 +54,15 @@ public class InputType implements Serializable {
     public InputType() {
     }
 
-    public InputType(Integer id) {
+    public InputType(Short id) {
         this.id = id;
     }
 
-    public Integer getId() {
+    public Short getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Short id) {
         this.id = id;
     }
